+++ apps/app/src/test/scala/com/daml/network/integration/tests/SvOffboardingIntegrationTest.scala
@@ -74,6 +74,7 @@ class SvOffboardingIntegrationTest
       clue("Initialize SVC with 4 SVs") {
         startAllSync(
           sv1ScanBackend,
+          sv2ScanBackend,
           sv1Backend,
           sv2Backend,
           sv3Backend,