+++ apps/app/src/test/scala/com/daml/network/integration/tests/SvOnboardingAddlIntegrationTest.scala
@@ -22,6 +22,7 @@ class SvOnboardingAddlIntegrationTest extends SvIntegrationTestBase {
     clue("Initialize SVC with 3 SVs") {
       startAllSync(
         sv1ScanBackend,
+        sv2ScanBackend,
         sv1Backend,
         sv2Backend,
         sv3Backend,
@@ -285,6 +286,7 @@ class SvOnboardingAddlIntegrationTest extends SvIntegrationTestBase {
       clue("Initialize SVC with 2 SVs") {
         startAllSync(
           sv1ScanBackend,
+          sv2ScanBackend,
           sv1Backend,
           sv2Backend,
           sv1ValidatorBackend,