+++ apps/app/src/test/scala/com/daml/network/integration/tests/SvTimeBasedOnboardingIntegrationTest.scala
@@ -29,6 +29,7 @@ class SvTimeBasedOnboardingIntegrationTest
       clue("Initialize SVC with 3 SVs") {
         startAllSync(
           sv1ScanBackend,
+          sv2ScanBackend,
           sv1Backend,
           sv2Backend,
           sv3Backend,